Creacion de controller, clase y repository

+++ Apis/Modelo/Restaurante.cs
@@ -5,7 +5,7 @@
         public int Id { get; set; }
         public string Nombre { get; set; }
         public string Direccion { get; set; }
-        public int CostoComercial { get; set; }
+        public double CostoComercial { get; set; }
         public Boolean PetFriendly { get; set; }
 
     }
